Autoscript to install some of my apps I use.
Auto reinstall is mainly for myself but it basically just installs steam, discord, spotify, and firefox deb version.
If you get an error about discord not installing is because discord either changed their web address under wget https://dl.discordapp.net/apps/linux/0.0.18/discord-0.0.18.deb to a newer version. Example discord changed to version 19 to change the 0.0.18 to 18 and the other similar line as well. If you know how to automatically grab latest version than let me know on how to improve this script. Other than that the rest of the script should not have any issues in theory. Feel free to file a bug report under issues if a bug is found.